requirements.txt
fastapi
uvicorn
pytesseract
pillow
python-docx
python-multipart
